requirements.txt
Flask
Flask-RESTful
werkzeug
Pillow
uwsgi